Add files via upload
On the button click, the callnumber from the textbox is used as a parameter for the searchBooks function which adds the feedback to the database and returns the modified shelf number (*currently only works for concourse*) for the getShelfLoc function. This then returns an array of the data needed to display the rectangle. I don't know how to use this data to display the target rectangle.